# Add JMH benchmarks for Result vs Promise performance comparison