--- a/cluster/src/test/java/org/pragmatica/cluster/consensus/rabia/CrashStopFailuresTest.java
+++ b/cluster/src/test/java/org/pragmatica/cluster/consensus/rabia/CrashStopFailuresTest.java
@@ -7,6 +7,7 @@ import org.pragmatica.cluster.consensus.rabia.infrastructure.TestCluster.StringK
 import org.pragmatica.cluster.net.NodeId;
 import org.pragmatica.cluster.state.kvstore.KVCommand;
 import org.pragmatica.cluster.state.kvstore.KVStoreNotification;
+import org.pragmatica.message.MessageRouter;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -127,7 +128,7 @@ public class CrashStopFailuresTest {
         var crashLatch = new CountDownLatch(numCommands / 3);
 
         // Set an observer to crash the node after some commands
-        cluster.routers().get(nodeToFail).addRoute(KVStoreNotification.ValuePut.class, _ -> {
+        ((MessageRouter.MutableRouter) cluster.routers().get(nodeToFail)).addRoute(KVStoreNotification.ValuePut.class, _ -> {
             int count = counter.incrementAndGet();
             if (count == numCommands / 3) {
                 log.info("Crashing node {} after {} commands", nodeToFail, count);