--- a/cluster/src/test/java/org/pragmatica/cluster/consensus/rabia/infrastructure/TestTopologyManager.java
+++ b/cluster/src/test/java/org/pragmatica/cluster/consensus/rabia/infrastructure/TestTopologyManager.java
@@ -34,6 +34,6 @@ public record TestTopologyManager(int clusterSize, NodeInfo self) implements Top
     }
 
     @Override
-    public void configure(MessageRouter router) {
+    public void configure(MessageRouter.MutableRouter router) {
     }
 }