… core/src/test/java/org/pragmatica/lang/benchmark/PerformanceAnalysis.java
… core/src/test/java/org/pragmatica/lang/benchmark/ResultVsPromiseBenchmark.java
… docs/performance/Result-vs-Promise-Benchmark-Results.md